requirements.txt
beautifulsoup4==4.14.3
certifi==2026.2.25
charset-normalizer==3.4.4
et_xmlfile==2.0.0
idna==3.11
numpy==2.2.6
openpyxl==3.1.5
pandas==2.3.3
pandas-stubs==2.3.3.260113
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
requests==2.32.5
six==1.17.0
soupsieve==2.8.3
types-pytz==2025.2.0.20251108
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
